requirements.txt
confluent-kafka>=2.10.1
impyla==0.21.0
python-magic==0.4.27
python-dotenv==1.0.0
lxml>=6.0.0
pyarrow==20.0.0
psycopg2==2.9.10